refactor(validation) : 값 확인 변수 정리
