+++ src/main/java/baseball/Constant.java
@@ -4,4 +4,5 @@ public class Constant {
     static final int MIN_NUMBER = 1;
     static final int MAX_NUMBER = 9;
     static final int MAX_LENGTH = 3;
+    static final int GAME_LENGTH = 1;
 }
